# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/configuracion": {
      "get": {
        "description": "Datos para armar comprobantes. La configuración se edita solo desde el Facturador. Scope `comprobantes:read`.",
        "operationId": "V1Controller_configuracion",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionPublicaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Ver emisor y puntos de venta",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/items": {
      "post": {
        "description": "Ingresa los ítems al staging (scope `items:write`). Es idempotente por `referenciaExterna`: reenviar un ítem igual no lo duplica y uno distinto lo actualiza (si todavía no se facturó). Cada ítem vuelve con su estado y sus errores.",
        "operationId": "V1Controller_cargarItems",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opcional (obligatoria para emitir). 8 a 100 caracteres. Reintentar con la misma clave devuelve la misma respuesta durante 24 h.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargarItemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoCargaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Cargar ítems a facturar",
        "tags": [
          "API pública"
        ]
      },
      "get": {
        "description": "Hasta 200 referencias (`?referencia=a&referencia=b`). Scope `items:write`.",
        "operationId": "V1Controller_items",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referencia",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemPublicoDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Consultar ítems por tu referencia",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/importaciones/{id}": {
      "get": {
        "operationId": "V1Controller_importacion",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoCargaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Ver una importación y sus ítems",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/importaciones/{id}/confirmar": {
      "post": {
        "description": "Deja sus ítems válidos listos para generar borradores.",
        "operationId": "V1Controller_confirmar",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opcional (obligatoria para emitir). 8 a 100 caracteres. Reintentar con la misma clave devuelve la misma respuesta durante 24 h.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoCargaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Confirmar una importación",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/clientes": {
      "get": {
        "description": "Activos, por documento o nombre (`?q=`), hasta 200. Scope `items:write`.",
        "operationId": "V1Controller_buscarClientes",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientePublicoDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Listar o buscar clientes",
        "tags": [
          "API pública"
        ]
      },
      "post": {
        "description": "Valida el documento (dígito verificador del CUIT). `409` si ya existe. Con `referenciaExterna` queda asociado a tu id. Scope `items:write`.",
        "operationId": "V1Controller_crearCliente",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opcional (obligatoria para emitir). 8 a 100 caracteres. Reintentar con la misma clave devuelve la misma respuesta durante 24 h.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearClienteApiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientePublicoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Dar de alta un cliente",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/clientes/padron/{cuit}": {
      "get": {
        "description": "Razón social, domicilio y condición de IVA para dar de alta al cliente. No guarda nada. Scope `items:write`.",
        "operationId": "V1Controller_padron",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cuit",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PadronDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Consultar un CUIT en el padrón de ARCA",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/clientes/referencias": {
      "get": {
        "description": "Hasta 200 (`?referencia=a&referencia=b`). Solo vuelven los que ya están asociados. Scope `items:write`.",
        "operationId": "V1Controller_clientesPorReferencia",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referencia",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteVinculadoDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Clientes asociados a tus ids",
        "tags": [
          "API pública"
        ]
      }
    },
    "/api/v1/clientes/{id}/referencias": {
      "post": {
        "description": "Para usar un cliente que ya existe en el Facturador. Scope `items:write`.",
        "operationId": "V1Controller_vincularCliente",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VincularReferenciaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteVinculadoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Asociar un cliente a tu id",
        "tags": [
          "API pública"
        ]
      }
    },
...
    "/api/v1/comprobantes/{id}/verificar": {
      "post": {
        "description": "Consulta en ARCA un comprobante PENDIENTE_VERIFICACION (también se hace solo, cada pocos minutos). En otro estado lo devuelve igual. Scope `comprobantes:write`.",
        "operationId": "V1Controller_verificar",
        "parameters": [
          {
            "name": "X-Usuario-Email",
            "in": "header",
            "description": "Opcional: email de la persona que hace la acción en tu sistema. Queda en la auditoría del Facturador.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComprobantePublicoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Verificar un comprobante pendiente",
        "tags": [
          "API pública"
        ]
      }
    },
...
  },
  "info": {
    "title": "Facturador — API pública",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key de la integración (fct_…)."
      }
    },
    "schemas": {
...
      "ConfiguracionPublicaDto": {
        "type": "object",
        "properties": {
          "emisor": {
            "$ref": "#/components/schemas/EmisorPublicoDto"
          },
          "puntosVenta": {
            "description": "Puntos de venta activos del ambiente actual (para `puntoVentaId` al crear un borrador).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PuntoVentaPublicoDto"
            }
          }
        },
        "required": [
          "emisor",
          "puntosVenta"
        ]
      },
...
      "CargarItemsDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 1000,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemApiDto"
            }
          },
          "descripcion": {
            "type": "string",
            "description": "Texto para reconocer el lote en el Facturador.",
            "maxLength": 200
          },
          "confirmar": {
            "type": "boolean",
            "description": "Confirmar el lote en el mismo request (deja los ítems válidos listos para facturar)."
          }
        },
        "required": [
          "items"
        ]
      },
...
      "VincularReferenciaDto": {
        "type": "object",
        "properties": {
          "referenciaExterna": {
            "type": "string",
            "description": "Id del cliente en tu sistema. Si ya estaba asociado a otro cliente, pasa a este.",
            "maxLength": 200
          }
        },
        "required": [
          "referenciaExterna"
        ]
      },
...
    }
  }
}